Clean-up
In this commit i;

- Move components to component folder
- Add types
- General improvements
