+++ App.tsx
@@ -11,18 +11,15 @@ import {
   Text,
   View,
   NativeModules,
-  Modal,
-  SafeAreaView,
-  Button,
-  TouchableOpacity,
-  ActivityIndicator,
+  SafeAreaView, 
   Linking,
   Platform,
   ToastAndroid,
   FlatList,
   RefreshControl
 } from 'react-native';
-
+import AppModal from './components/AppModal'
+import LoaderComponent from './components/Loader'
 
 function App(): React.JSX.Element {
   const [showPermissionModal, setShowPermissionModal] = useState(false)
@@ -129,43 +126,6 @@ function App(): React.JSX.Element {
   );
 }
 
-const LoaderComponent = ({ isLoading }: { isLoading: boolean }) => {
-  return (
-    <>
-      {
-        isLoading ?
-          <ActivityIndicator color={'green'} /> : null
-      }
-    </>
-  )
-}
-
-const AppModal = ({ showPermissionModal, onButtonPressed }: { showPermissionModal: boolean, onButtonPressed: () => void }) => {
-  return (
-    <>
-      <Modal
-        visible={showPermissionModal}
-        animationType='slide'
-        statusBarTranslucent
-        transparent
-      >
-        <SafeAreaView style={styles.modalContainer}>
-          <View style={styles.modalContentContainer}>
-            <Text style={styles.modalContentViewTitle}>Grant Permission</Text>
-            <Text style={styles.modalContentViewSubTitle}>You need to grant permission to HealthApp before you can start enjoying our services. We will never share your data with third party</Text>
-            <TouchableOpacity
-              onPress={() => onButtonPressed()}
-              style={styles.button}
-            >
-              <Text>Request Permission</Text>
-            </TouchableOpacity>
-          </View>
-        </SafeAreaView>
-      </Modal>
-    </>
-  )
-}
-
 const EmptyStateMessage = () => {
   return (
     <Text style={styles.emptyStateMessage}>No health data available</Text>
@@ -175,37 +135,6 @@ const EmptyStateMessage = () => {
 const styles = StyleSheet.create({
   containter: {
     flex: 1,
-  },
-  modalContainer: {
-    flex: 1,
-    alignContent: 'center',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'black',
-    opacity: .9
-  },
-  modalContentContainer: {
-    flex: 1,
-    alignContent: 'center',
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: 20
-  },
-  modalContentViewTitle: {
-    color: 'white',
-    fontSize: 25,
-    marginBottom: 15
-  },
-  modalContentViewSubTitle: {
-    color: 'white',
-    fontSize: 15,
-    textAlign: "justify",
-    marginBottom: 15
-  },
-  button: {
-    alignItems: 'center',
-    backgroundColor: '#DDDDDD',
-    padding: 10
   },
   dataItem: {
     flexDirection: 'row',
